+++ tests/UnitTests/Curiosity.Migrations.UnitTests/CodeMigrations/ThirdMigration.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 
 namespace Curiosity.Migrations.UnitTests.CodeMigrations
 {
-    public class ThirdMigration : CodeMigration, ISpecificCodeMigrations
+    public class ThirdMigration : CodeMigration, ISpecificCodeMigrations, IDowngradeMigration
     {
         public override DbVersion Version { get; } = new DbVersion(1,2);
         
@@ -15,7 +15,7 @@ namespace Curiosity.Migrations.UnitTests.CodeMigrations
             return DbProvider.ExecuteScriptAsync(ScriptConstants.UpScript, token);
         }
 
-        public override Task DowngradeAsync(DbTransaction transaction, CancellationToken token = default)
+        public Task DowngradeAsync(DbTransaction transaction, CancellationToken token = default)
         {
             return DbProvider.ExecuteScriptAsync(ScriptConstants.DownScript, token);
         }