+++ tests/UnitTests/Curiosity.Migrations.UnitTests/ScriptMigrationsProvider_Should.cs
@@ -31,29 +31,27 @@ namespace Curiosity.Migrations.UnitTests
             
             Assert.Equal(5, migrations.Count);
             
-            Assert.True(migrations[0] is ScriptMigration);
+            Assert.True(migrations[0] is DowngradableScriptMigration);
             Assert.Equal(new DbVersion(1,0), migrations[0].Version);
             Assert.Equal("comment", migrations[0].Comment);
-            Assert.Equal("up", ((ScriptMigration)migrations[0]).UpScripts[0].Script);
-            Assert.Equal("down", ((ScriptMigration)migrations[0]).DownScripts[0].Script);
+            Assert.Equal("up", ((DowngradableScriptMigration)migrations[0]).UpScripts[0].Script);
+            Assert.Equal("down", ((DowngradableScriptMigration)migrations[0]).DownScripts[0].Script);
             
-            Assert.True(migrations[1] is ScriptMigration);
+            Assert.True(migrations[1] is DowngradableScriptMigration);
             Assert.Equal(new DbVersion(1,1), migrations[1].Version);
             Assert.True(String.IsNullOrEmpty(migrations[1].Comment));
-            Assert.Equal("up", ((ScriptMigration)migrations[1]).UpScripts[0].Script);
-            Assert.Equal("down", ((ScriptMigration)migrations[1]).DownScripts[0].Script);
+            Assert.Equal("up", ((DowngradableScriptMigration)migrations[1]).UpScripts[0].Script);
+            Assert.Equal("down", ((DowngradableScriptMigration)migrations[1]).DownScripts[0].Script);
 
             Assert.True(migrations[2] is ScriptMigration);
             Assert.Equal(new DbVersion(1,2), migrations[2].Version);
             Assert.Equal("comment", migrations[2].Comment);
             Assert.Equal("up", ((ScriptMigration)migrations[2]).UpScripts[0].Script);
-            Assert.True(((ScriptMigration)migrations[2]).DownScripts.Count == 0);
             
             Assert.True(migrations[3] is ScriptMigration);
             Assert.Equal(new DbVersion(1,3), migrations[3].Version);
             Assert.True(String.IsNullOrEmpty(migrations[3].Comment));
             Assert.Equal("up", ((ScriptMigration)migrations[3]).UpScripts[0].Script);
-            Assert.True(((ScriptMigration)migrations[3]).DownScripts.Count == 0);
         }
         
         [Fact]
@@ -76,7 +74,6 @@ namespace Curiosity.Migrations.UnitTests
             Assert.Equal(new DbVersion(0,1), migrations[0].Version);
             Assert.True(String.IsNullOrEmpty(migrations[0].Comment));
             Assert.Equal("prefix", ((ScriptMigration)migrations[0]).UpScripts[0].Script);
-            Assert.True(((ScriptMigration)migrations[0]).DownScripts.Count == 0);
         }
         
         [Fact]
@@ -94,38 +91,35 @@ namespace Curiosity.Migrations.UnitTests
             
             Assert.Equal(5, migrations.Count);
             
-            Assert.True(migrations[0] is ScriptMigration);
+            Assert.True(migrations[0] is DowngradableScriptMigration);
             Assert.Equal(new DbVersion(1,0), migrations[0].Version);
             Assert.Equal("comment", migrations[0].Comment);
-            Assert.Equal("up", ((ScriptMigration)migrations[0]).UpScripts[0].Script);
-            Assert.Equal("down", ((ScriptMigration)migrations[0]).DownScripts[0].Script);
+            Assert.Equal("up", ((DowngradableScriptMigration)migrations[0]).UpScripts[0].Script);
+            Assert.Equal("down", ((DowngradableScriptMigration)migrations[0]).DownScripts[0].Script);
             
             
-            Assert.True(migrations[1] is ScriptMigration);
+            Assert.True(migrations[1] is DowngradableScriptMigration);
             Assert.Equal(new DbVersion(1,1), migrations[1].Version);
             Assert.True(String.IsNullOrEmpty(migrations[1].Comment));
-            Assert.Equal("up", ((ScriptMigration)migrations[1]).UpScripts[0].Script);
-            Assert.Equal("down", ((ScriptMigration)migrations[1]).DownScripts[0].Script);
+            Assert.Equal("up", ((DowngradableScriptMigration)migrations[1]).UpScripts[0].Script);
+            Assert.Equal("down", ((DowngradableScriptMigration)migrations[1]).DownScripts[0].Script);
             
             
             Assert.True(migrations[2] is ScriptMigration);
             Assert.Equal(new DbVersion(1,2), migrations[2].Version);
             Assert.Equal("comment", migrations[2].Comment);
             Assert.Equal("up", ((ScriptMigration)migrations[2]).UpScripts[0].Script);
-            Assert.True(((ScriptMigration)migrations[2]).DownScripts.Count == 0);
             
             Assert.True(migrations[3] is ScriptMigration);
             Assert.Equal(new DbVersion(1,3), migrations[3].Version);
             Assert.True(String.IsNullOrEmpty(migrations[3].Comment));
             Assert.Equal("up", ((ScriptMigration)migrations[3]).UpScripts[0].Script);
-            Assert.True(((ScriptMigration)migrations[3]).DownScripts.Count == 0);
             
             
             Assert.True(migrations[4] is ScriptMigration);
             Assert.Equal(new DbVersion(2,0), migrations[4].Version);
             Assert.True(String.IsNullOrEmpty(migrations[4].Comment));
             Assert.Equal("prefix", ((ScriptMigration)migrations[4]).UpScripts[0].Script);
-            Assert.True(((ScriptMigration)migrations[4]).DownScripts.Count == 0);
         }
         
         [Fact]
@@ -147,7 +141,6 @@ namespace Curiosity.Migrations.UnitTests
             Assert.Equal(new DbVersion(2,0), migrations[0].Version);
             Assert.True(String.IsNullOrEmpty(migrations[0].Comment));
             Assert.Equal("prefix", ((ScriptMigration)migrations[0]).UpScripts[0].Script);
-            Assert.True(((ScriptMigration)migrations[0]).DownScripts.Count == 0);
         }
         
         [Fact]