+++ src/Curiosity.Migrations/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [2.0] - 2020-05-14
+
+### Changed
+- Splitted downgrade migrations into another interface for safety reasons.
+- Default `downgrade policy` is `Forbidden`
+
 ## [1.2.0] - 2020-04-23
 
 ### Added